Add logging, make fact throw exceptions if the input number is to big

+++ src/factorial/factorial.h
@@ -1,6 +1,6 @@
 #pragma once
-
+#include <cstddef> // for size_t
 namespace factorial
 {
-    [[nodiscard]] int computeFact(int factorialNumber);
+    [[nodiscard]] size_t computeFact(size_t factorialNumber);
 }
